修改订单api，分页，按最近时间排序； 修改权限菜单api，保证菜单的删除添加，同步关联角色； 商品列表api，增加搜索功能；

+++ routes/admin/menu.js
@@ -17,19 +17,30 @@ let db = require('../../config/mysql');
  *
  * @apiSampleRequest /api/menu
  */
-router.post("/", function (req, res) {
-    let {name, pId, component, path, order} = req.body;
-    let sql = `INSERT INTO MENU (name,pId,component,path,menu_order) VALUES (?,?,?,?,?) `;
-    db.query(sql, [name, pId, component, path, order], function (results, fields) {
-        //成功
-        res.json({
-            status: true,
-            msg: "success!",
-            data: {
-                id: results.insertId
-            }
-        });
-    });
+router.post("/", function(req, res) {
+	let { name, pId, component, path, order } = req.body;
+	let sql = `INSERT INTO MENU (name,pId,component,path,menu_order) VALUES (?,?,?,?,?) `;
+	db.query(sql, [name, pId, component, path, order], function({ insertId }, fields) {
+		let sql = `INSERT INTO role_menu (role_id,menu_id) VALUES (1,?)`;
+		db.query(sql, [insertId], function(results, fields) {
+			if (!results.affectedRows) {
+				res.json({
+					status: false,
+					msg: "创建失败！"
+				});
+				return;
+			}
+			//成功
+			res.json({
+				status: true,
+				msg: "创建成功!",
+				data: {
+					id: insertId
+				}
+			});
+		})
+
+	});
 });
 /**
  * @api {delete} /api/menu 删除子菜单
@@ -41,16 +52,23 @@ router.post("/", function (req, res) {
  *
  * @apiSampleRequest /api/menu
  */
-router.delete("/", function (req, res) {
-    let {id} = req.query;
-    let sql = `DELETE FROM MENU WHERE id = ?`;
-    db.query(sql, [id], function (results, fields) {
-        //成功
-        res.json({
-            status: true,
-            msg: "success!"
-        });
-    });
+router.delete("/", function(req, res) {
+	let { id } = req.query;
+	let sql = `DELETE FROM MENU WHERE id = ?;DELETE FROM ROLE_MENU WHERE menu_id = ?`;
+	db.query(sql, [id, id], function(results, fields) {
+		if (!results[1].affectedRows) {
+			res.json({
+				status: false,
+				msg: "删除失败！"
+			});
+			return;
+		}
+		//成功
+		res.json({
+			status: true,
+			msg: "删除成功!"
+		});
+	});
 });
 /**
  * @api {put} /api/menu 更新子菜单
@@ -67,16 +85,23 @@ router.delete("/", function (req, res) {
  *
  * @apiSampleRequest /api/menu
  */
-router.put("/", function (req, res) {
-    let {id, name, pId, component, path, order} = req.body;
-    let sql = `UPDATE MENU SET name = ?,pId = ?,component = ?, path = ?, menu_order = ? WHERE id = ? `;
-    db.query(sql, [name, pId, component, path, order, id], function (results, fields) {
-        //成功
-        res.json({
-            status: true,
-            msg: "success!"
-        });
-    });
+router.put("/", function(req, res) {
+	let { id, name, pId, component, path, order } = req.body;
+	let sql = `UPDATE MENU SET name = ?,pId = ?,component = ?, path = ?, menu_order = ? WHERE id = ? `;
+	db.query(sql, [name, pId, component, path, order, id], function(results, fields) {
+		if (!results.affectedRows) {
+			res.json({
+				status: false,
+				msg: "failed！"
+			});
+			return;
+		}
+		//成功
+		res.json({
+			status: true,
+			msg: "success!"
+		});
+	});
 });
 /**
  * @api {put} /api/menu/icon 设置子菜单图标
@@ -89,16 +114,23 @@ router.put("/", function (req, res) {
  *
  * @apiSampleRequest /api/menu/icon
  */
-router.put("/icon", function (req, res) {
-    let {id, icon} = req.body;
-    let sql = `UPDATE MENU SET icon_id = ? WHERE id = ? `;
-    db.query(sql, [icon, id], function (results, fields) {
-        //成功
-        res.json({
-            status: true,
-            msg: "success!"
-        });
-    });
+router.put("/icon", function(req, res) {
+	let { id, icon } = req.body;
+	let sql = `UPDATE MENU SET icon_id = ? WHERE id = ? `;
+	db.query(sql, [icon, id], function(results, fields) {
+		if (!results.affectedRows) {
+			res.json({
+				status: false,
+				msg: "failed！"
+			});
+			return;
+		}
+		//成功
+		res.json({
+			status: true,
+			msg: "success!"
+		});
+	});
 });
 /**
  * @api {get} /api/menu/sub 获取子菜单
@@ -110,17 +142,17 @@ router.put("/icon", function (req, res) {
  *
  * @apiSampleRequest /api/menu/sub
  */
-router.get("/sub", function (req, res) {
-    let sql =
-        `SELECT m.id, m.name, m.pId, m.path, m.component, m.menu_order AS 'order', i.name AS 'icon' FROM MENU m LEFT JOIN ICON i ON m.icon_id = i.id WHERE m.pId = ? ORDER BY m.menu_order`;
-    db.query(sql, [req.query.pId], function (results, fields) {
-        //成功
-        res.json({
-            status: true,
-            msg: "success!",
-            data: results
-        });
-    });
+router.get("/sub", function(req, res) {
+	let sql =
+		`SELECT m.id, m.name, m.pId, m.path, m.component, m.menu_order AS 'order', i.name AS 'icon' FROM MENU m LEFT JOIN ICON i ON m.icon_id = i.id WHERE m.pId = ? ORDER BY m.menu_order`;
+	db.query(sql, [req.query.pId], function(results, fields) {
+		//成功
+		res.json({
+			status: true,
+			msg: "success!",
+			data: results
+		});
+	});
 });
 /**
  * @api {get} /api/menu/tree 根据角色id获取侧边栏树形菜单
@@ -132,33 +164,33 @@ router.get("/sub", function (req, res) {
  *
  * @apiSampleRequest /api/menu/tree
  */
-router.get('/tree', function (req, res) {
-    let {id} = req.query;
-    let sql =
-        `SELECT m.*,i.name AS 'icon' FROM MENU m JOIN role_menu rm ON rm.menu_id = m.id LEFT JOIN ICON i ON m.icon_id = i.id WHERE rm.role_id = ? ORDER BY menu_order;`;
-    db.query(sql, [id], function (results) {
-        //筛选出一级菜单
-        let cate_1st = results.filter((item) => item.pId === 1 ? item : null);
-        //递归循环数据
-        parseToTree(cate_1st);
-        //递归函数
-        function parseToTree(array) {
-            array.forEach(function(parent) {
-                parent.children = [];
-                results.forEach(function(child) {
-                    if (child.pId === parent.id) {
-                        parent.children.push(child);
-                    }
-                });
-                parseToTree(parent.children);
-            });
-        }
-        //成功
-        res.json({
-            status: true,
-            msg: "success!",
-            data: cate_1st
-        });
-    });
+router.get('/tree', function(req, res) {
+	let { id } = req.query;
+	let sql =
+		`SELECT m.*,i.name AS 'icon' FROM MENU m JOIN role_menu rm ON rm.menu_id = m.id LEFT JOIN ICON i ON m.icon_id = i.id WHERE rm.role_id = ? ORDER BY m.menu_order;`;
+	db.query(sql, [id], function(results) {
+		//筛选出一级菜单
+		let cate_1st = results.filter((item) => item.pId === 1 ? item : null);
+		//递归循环数据
+		parseToTree(cate_1st);
+		//递归函数
+		function parseToTree(array) {
+			array.forEach(function(parent) {
+				parent.children = [];
+				results.forEach(function(child) {
+					if (child.pId === parent.id) {
+						parent.children.push(child);
+					}
+				});
+				parseToTree(parent.children);
+			});
+		}
+		//成功
+		res.json({
+			status: true,
+			msg: "success!",
+			data: cate_1st
+		});
+	});
 });
 module.exports = router;
